moving logic to hook
